fix skill dates transformation
